fix: cast int to double before calling toStringAsFixed in _sanitizeCoordinate
int does not have toStringAsFixed() in Dart, causing a runtime
NoSuchMethodError. Added .toDouble() cast before calling toStringAsFixed(6).

Closes #4169

+++ apps/driver/lib/screens/home_screen.dart
@@ -154,7 +154,7 @@ class _HomeScreenState extends State<HomeScreen> {
   String _sanitizeCoordinate(dynamic coord) {
     if (coord == null) return '0.0';
     if (coord is double) return coord.toStringAsFixed(6);
-    if (coord is int) return coord.toStringAsFixed(6);
+    if (coord is int) return coord.toDouble().toStringAsFixed(6);
     return (double.tryParse(coord.toString()) ?? 0.0).toStringAsFixed(6);
   }
